Start non-touch theme, set menubar padding
Eyeballed menubar height from 6th ed emulator and matched. Note that the bottom
padding on the MSK is 3 px bigger than LSK or RSK. This is on purpose. If they
are the same, and there is only one command on MSK and none on LSK and RSK, the
menubar becomes too low.

Change-Id: Ib63e375d9e641ebad06424754c12a229c9e0e5ab