Reset menubar buttons to unpressed when dragging away from them
Fixes: sid-24749713

Whenever a pointer press happens on the menubar, store the component
that was the target, because if it is a button, it will go in the pressed
state. Now when the finger is not released but instead dragged away,
use this information to reset the button to unpressed when the finger
is lifted, without triggering the associated action. For this we add
a new public API resetState() to Button.

Change-Id: I4729974201126770554490939433b62dcd01fc64