Implement support for S40 IconCommands
With these changes, LWUIT Commands that have icons defined get turned into native IconCommands. This means that if they are displayed on the FT action button, the developer can show a custom icon.

Note that the built-in system icon IDs cannot be used in this solution, the icons need to be defined on the LWUIT Commands as LWUIT Images.

Change-Id: I573853070c376dd13d9d47dddf30d8591cffc561